+++ lib/phx_bb_web/live/components/breadcrumb_component.ex
@@ -7,12 +7,22 @@ defmodule PhxBbWeb.BreadcrumbComponent do
 
   import PhxBbWeb.StyleHelpers
 
-  def mount(socket) do
-    {:ok, socket}
-  end
-
-  def update(assigns, socket) do
-    socket = assign(socket, assigns)
-    {:ok, socket}
+  def render(assigns) do
+    ~L"""
+    <div class="pt-4 pb-4 md:pl-8 md:pt-8" id="breadcrumb">
+       >
+      <%= live_patch "Board Index",
+            to: Routes.live_path(@socket, PhxBbWeb.PageLive),
+            class: link_style(@active_user),
+            id: "crumb-index" %>
+      <%= if @nav in [:post, :create_post] do %>
+       >
+      <%= live_patch @active_board.name,
+            to: Routes.live_path(@socket, PhxBbWeb.PageLive, board: @active_board.id),
+            class: link_style(@active_user),
+            id: "crumb-board" %>
+      <% end %>
+    </div>
+    """
   end
 end