+++ CHANGELOG.md
@@ -6,6 +6,13 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
 
+### Added
+
+- Administrator account: can edit and delete posts
+- Users can now edit and delete their own posts
+
+## [0.4.1] - 2021-05-04
+
 ### Fixed
 
 - Markdown is now properly parsed in user profiles' last 5 posts
@@ -104,7 +111,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - User cache for reduced DB queries
 - Page title updates to reflect current view
 
-[Unreleased]: https://github.com/APB9785/phxBB/compare/0.4...HEAD
+[Unreleased]: https://github.com/APB9785/phxBB/compare/0.4.1...HEAD
+[0.4.1]: https://github.com/APB9785/phxBB/compare/0.4...0.4.1
 [0.4]: https://github.com/APB9785/phxBB/compare/0.3.1...0.4
 [0.3.1]: https://github.com/APB9785/phxBB/compare/0.3...0.3.1
 [0.3]: https://github.com/APB9785/phxBB/compare/0.2...0.3